+++ src/graphql-api/fragments.ts
@@ -366,10 +366,7 @@ export const PRIVATE_CHAT_FRAGMENT = `
   ... on PrivateChat {
     __typename
     id
-    messages(
-      last: $privateChat_messages_last
-      before: $privateChat_messages_before
-    ) {
+    messages(last: $privateChat_messages_last, before: $privateChat_messages_before) {
       ${MESSAGES_CONNECTION_FRAGMENT}
     }
     user {
@@ -779,10 +776,7 @@ export const GROUP_CHAT_FRAGMENT = `
     users(first: $groupChat_users_first, after: $groupChat_users_after) {
       ${ACCOUNTS_CONNECTION_FRAGMENT}
     }
-    messages(
-      last: $groupChat_messages_last
-      before: $groupChat_messages_before
-    ) {
+    messages(last: $groupChat_messages_last, before: $groupChat_messages_before) {
       ${MESSAGES_CONNECTION_FRAGMENT}
     }
     isBroadcast