+++ src/graphql-api/operator.ts
@@ -18,12 +18,8 @@ import {
   UnverifiedEmailAddressError,
   UsernameTakenError,
 } from './errors';
-import {
-  ConnectionError,
-  InternalServerError,
-  UnauthorizedError,
-} from '../errors';
-import {ApiUrl, HttpProtocol, WebSocketProtocol} from '../config';
+import { ConnectionError, InternalServerError, UnauthorizedError } from '../errors';
+import { ApiUrl, HttpProtocol, WebSocketProtocol } from '../config';
 
 export interface GraphQlRequest {
   readonly query: string;
@@ -73,19 +69,17 @@ export async function queryOrMutate(
   protocol: HttpProtocol,
   apiUrl: ApiUrl,
   request: GraphQlRequest,
-  accessToken?: string
+  accessToken?: string,
 ): Promise<GraphQlResponse> {
-  const headers: Record<string, string> = {'Content-Type': 'application/json'};
-  if (accessToken !== undefined)
-    headers.Authorization = `Bearer ${accessToken}`;
+  const headers: Record<string, string> = { 'Content-Type': 'application/json' };
+  if (accessToken !== undefined) headers.Authorization = `Bearer ${accessToken}`;
   const response = await fetch(`${protocol}://${apiUrl}/query-or-mutation`, {
     method: 'POST',
     headers,
     body: JSON.stringify(request),
   });
   if (response.status === 401) throw new UnauthorizedError();
-  if (response.status >= 500 && response.status <= 599)
-    throw new InternalServerError();
+  if (response.status >= 500 && response.status <= 599) throw new InternalServerError();
   if (response.status !== 200) throw new ConnectionError();
   const graphQlResponse = (await response.json()) as GraphQlResponse;
   handleGraphQlError(graphQlResponse.errors);
@@ -93,9 +87,8 @@ export async function queryOrMutate(
 }
 
 /**
- * Executed when the socket closes due to an unexpected error. An example use
- * case is displaying a message notifying the user that the server is currently
- * unreachable.
+ * Executed when the socket closes due to an unexpected error. An example use case is displaying a message notifying the
+ * user that the server is currently unreachable.
  */
 export interface OnSocketError {
   (): void;
@@ -137,11 +130,9 @@ export interface OnSocketClose {
 /**
  * Creates a GraphQL subscription.
  * @param operation Example: `'subscribeToAccounts'`.
- * @param path For example, if the subscription is hosted on
- * http://localhost/accounts-subscription, this should be
+ * @param path For example, if the subscription is hosted on http://localhost/accounts-subscription, this should be
  * `/accounts-subscription`.
- * @param query GraphQL document (i.e., the query to send to the GraphQL
- * server).
+ * @param query GraphQL document (i.e., the query to send to the GraphQL server).
  */
 export function subscribe<T>(
   protocol: WebSocketProtocol,
@@ -151,14 +142,14 @@ export function subscribe<T>(
   path: string,
   query: string,
   onMessage: OnSocketMessage<T>,
-  onError: OnSocketError
+  onError: OnSocketError,
 ): OnSocketClose {
   const socket = new WebSocket(`${protocol}://${apiUrl}${path}`);
   socket.addEventListener('open', () => {
     socket.send(accessToken);
-    socket.send(JSON.stringify({query}));
+    socket.send(JSON.stringify({ query }));
   });
-  socket.addEventListener('message', ({data}) => {
+  socket.addEventListener('message', ({ data }) => {
     const response = JSON.parse(data) as GraphQlResponse;
     if (response.errors !== undefined) {
       onError();
@@ -170,13 +161,11 @@ export function subscribe<T>(
 }
 
 /**
- * Throws the relevant error if it's documented in the API's docs. For example,
- * if `errors[0].message` is `'NONEXISTENT_USER'`, a
- * {@link NonexistentUserError} will be thrown. Otherwise, a
- * {@link ConnectionError} will be thrown.
- * @param errors No error is thrown if this is `undefined`. If this is a
- * `GraphQLError[]`, it's assumed at least one element is present as per the
- * GraphQL spec.
+ * Throws the relevant error if it's documented in the API's docs. For example, if `errors[0].message` is
+ * `'NONEXISTENT_USER'`, a {@link NonexistentUserError} will be thrown. Otherwise, a {@link ConnectionError} will be
+ * thrown.
+ * @param errors No error is thrown if this is `undefined`. If this is a `GraphQLError[]`, it's assumed at least one
+ * element is present as per the GraphQL spec.
  * @throws {@link InternalServerError}
  * @throws {@link NonexistentUserError}
  * @throws {@link UnverifiedEmailAddressError}